Fix web-typecheck CI: bun mock.module / env leaks across test files
The new push-mute route test ran fine alone but broke other files under CI's
single-process `bun test`: it mutated process-global `process.env.VERCEL` and
`mock.module(\"../services/apns/sender\")`, which leaked into
notifications-push-route.test.ts (rate-limit-first expectation) and apns.test.ts
(real-sender transport tests + signApnsJwt import).

Merge the mute tests into notifications-push-route.test.ts (one file, one shared
mock set, no cross-file leak): the rate limiter is allowed through per-test via
the existing checkRateLimit mock, cloudDb delegates to a swappable factory, and
the sender is not mocked at all (the table-access assertions already prove the
muted path never reaches a send). Removed the redundant separate test file.

web tsc/lint/full `bun test` green (122 pass).
